% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_estimates.R
\name{adjust_estimates}
\alias{adjust_estimates}
\title{Adjust the estimates to account for correlated coefficients in two-sample MVMR.}
\usage{
adjust_estimates(
  coef_x,
  se_coef_x,
  coef_y,
  se_coef_y,
  cor_zinx,
  cor_ziny,
  cor_x,
  nobs_x,
  nobs_y,
  ridge_pen_zinx = 0,
  ridge_pen_ziny = 0
)
}
\arguments{
\item{coef_x}{A matrix of dimension J x K, the simple OLS coefficients of
each X on each Z.}

\item{coef_y}{A vector of length J, the simple OLS coefficients of Y on each Z.}

\item{cor_zinx}{A matrix of dimension J x J, the correlation matrix of the instruments in sample with X (LD matrix).}

\item{cor_ziny}{A matrix of dimension J x J, the correlation matrix of the instruments in sample with Y (LD matrix).}

\item{cor_x}{A matrix of dimension K x K, the correlation matrix of X.}

\item{nobs_x}{The number of observations used to estimate \code{coef_x}.}

\item{nobs_y}{The number of observations used to estimated \code{coef_y}.}

\item{cov_coef_x}{A matrix of dimension J x J, the covariance matrix of \code{coef_x}.}

\item{cov_coef_y}{A vector of length J, the variances of \code{coef_y}.}
}
\value{
A list containing
}
\description{
Adjust the estimates to account for correlated coefficients in two-sample MVMR.
}
\references{
Wang and Kang 2021 for the univariate X case and Patel et al 2023
for multivariate X case.
}
