% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_estimates.R
\name{adjust_estimates_zx}
\alias{adjust_estimates_zx}
\title{Adjust sufficient statistics to account for correlations}
\usage{
adjust_estimates_zx(coef_x, se_coef_x, cor_zinx, cor_x, nobs_x)
}
\description{
Adjust sufficient statistics to account for correlations
}
