% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_estimates.R
\name{adjust_estimates_zy}
\alias{adjust_estimates_zy}
\title{Adjust sufficient statistics to account for correlations}
\usage{
adjust_estimates_zy(coef_y, se_coef_y, cor_ziny, nobs_y)
}
\description{
Adjust sufficient statistics to account for correlations
}
