% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{anderson_rubin_test}
\alias{anderson_rubin_test}
\title{Anderson-Rubin test}
\usage{
anderson_rubin_test(X, Y, Z, beta, chisq_df = NULL)
}
\arguments{
\item{beta}{A vector of length K, the exposure-outcome coefficients.}
}
\value{
A list containing the test statistic \code{stat} and the corresponding
p-value \code{pval}.
}
\description{
Anderson-Rubin test
}
