% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{anderson_rubin_test_suffstat}
\alias{anderson_rubin_test_suffstat}
\title{Anderson-Rubin test with summary statistics}
\usage{
anderson_rubin_test_suffstat(
  XX,
  YY,
  ZZ,
  XY,
  ZX,
  ZY,
  nobs,
  beta,
  chisq_df = NULL
)
}
\arguments{
\item{XX}{A matrix of dimension R x R, the covariance matrix of the exposures.}

\item{YY}{A scalar, the variance of the response.}

\item{ZZ}{A matrix of dimension J x J, the covariance matrix of the instruments.}

\item{XY}{A vector of length R, the covariances between the exposures and the response.}

\item{ZX}{A matrix of dimension J x R, the covariances between the instruments and the exposures.}

\item{ZY}{A vector of length J, the covariances between the instruments and the response.}

\item{nobs}{A scalar, number of observations.}

\item{beta}{A vector of length K, the exposure-outcome coefficients.}
}
\value{
A list containing the test statistic \code{stat} and the corresponding
p-value \code{pval}.
}
\description{
Anderson-Rubin test with summary statistics
}
