% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{check_assumptions_A}
\alias{check_assumptions_A}
\title{Check if Assumptions A are satisfied}
\usage{
check_assumptions_A(A, B, parents, beta_star)
}
\arguments{
\item{A}{A matrix of dimension nins x nexp.}

\item{B}{A matrix of dimension nexp x nexp.}

\item{beta_star}{A vector of length nexp, the true exposure-response
coefficients.}
}
\value{
A list containing three elements, corresponding to the three
conditions in Assumptions A in Pfister and Peters 2021.
}
\description{
Check if Assumptions A are satisfied
}
