% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{comp_suffstat}
\alias{comp_suffstat}
\title{Calculate sufficient statistics given individual level data}
\usage{
comp_suffstat(Y, Z_Y, X, Z_X)
}
\arguments{
\item{Y}{A matrix of dimension nobs x 1, the individual observations of the
response.}

\item{Z_Y}{A matrix of dimension nobs x J, the individual observations of the
instruments associated with the response Y.}

\item{X}{A matrix of dimension nobs x K, the individual observations of the
exposures.}

\item{Z_X}{A matrix of dimension nobs x J, the individual observations of the
instruments associated with the exposures X.}
}
\value{
A list containing the variance-covariance matrices.
}
\description{
Calculate sufficient statistics given individual level data
}
