% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{comp_sumstat}
\alias{comp_sumstat}
\title{Compute summary statistics from two non-overlapping individual level samples.}
\usage{
comp_sumstat(Y, Z_Y, X, Z_X, type = c("depen_mv", "indep_mv", "simple"))
}
\arguments{
\item{Y}{A matrix of dimension nobs x 1, the individual observations of the
response.}

\item{Z_Y}{A matrix of dimension nobs x J, the individual observations of the
instruments associated with the response Y.}

\item{X}{A matrix of dimension nobs x K, the individual observations of the
exposures.}

\item{Z_X}{A matrix of dimension nobs x J, the individual observations of the
instruments associated with the exposures X.}

\item{complete}{A boolean indicating whether to calculate the full variance-covariance
matrix of the coefficients.}
}
\value{
A list containing
\itemize{
\item coef_x: A matrix of dimension J x K, the estimated instrument-exposure coefficients.
\item coef_y: A vector of length J, the estimated instrument-response coefficients.
\item cov_coef_x: if indep, a list of length K of matrices each of dimension J x J,
the covariance matrices of the estimated instrument-exposure coefficients;
if not indep, a matrix of dimension (J x K) x (J x K)
\item cov_coef_y: if indep, a vector of length J, the standard errors of the
estimated instrument-response coefficients; if not indep, a matrix of
dimension J x J, the covariance matrix of the estimated instrument-response
coefficients.
\item cor_x: A matrix of dimension K x K, the correlation matrix of X.
\item cor_z: A matrix of dimension J x J, the correlation matrix of Z.
}
}
\description{
Compute summary statistics from two non-overlapping individual level samples.
}
