% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{coordinate_descent_iv2_estimator}
\alias{coordinate_descent_iv2_estimator}
\title{Minimizing ||Cov(Z, Y-Xb)||^2 using coordinate descent}
\usage{
coordinate_descent_iv2_estimator(
  beta_init,
  X,
  Y,
  Z,
  max_iter = 500,
  abs_tol = 1e-05,
  verbose = FALSE
)
}
\description{
Minimizing ||Cov(Z, Y-Xb)||^2 using coordinate descent
}
\details{
IV solutions using valid instruments satisfy Cov(Z, Y-Xb) = 0.
The objective here is to minimize the squared LHS.
}
\examples{
dat <- dgp_ex00(1e3, c(1, 2)) # just-identified case
coordinate_descent_iv2_estimator(c(1, 1), dat$X, dat$Y, dat$Z, verbose = FALSE)
dat <- dgp_ex0(1e3, c(1, 2)) # over-identified case
coordinate_descent_iv2_estimator(c(1, 1), dat$X, dat$Y, dat$Z, verbose = FALSE)
dat <- dgp_ex1(1e3, 2) # under-identified case
coordinate_descent_iv2_estimator(c(1, 1, 1), dat$X, dat$Y, dat$Z, verbose = FALSE)
coordinate_descent_iv2_estimator(c(0, 2, 0), dat$X, dat$Y, dat$Z, verbose = FALSE)

}
