% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{coordinate_descent_iv_lasso}
\alias{coordinate_descent_iv_lasso}
\title{Coordinate descent with soft-thresholding for IV-Lasso}
\usage{
coordinate_descent_iv_lasso(
  beta_init,
  X,
  Y,
  Z,
  pen = 0.01,
  max_iter = 500,
  abs_tol = 1e-05,
  verbose = FALSE
)
}
\description{
Coordinate descent with soft-thresholding for IV-Lasso
}
