% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp1}
\alias{dgp1}
\title{Data generating process in Example 4 of spaceIV}
\usage{
dgp1(n, beta1, beta2, seed_num = NULL)
}
\description{
Data generating process in Example 4 of spaceIV
}
\details{
Example 4 in spaceIV with three hidden variables each affecting one
exposure as well as the outcome. A DGP with 3 instruments and 2 exposures.
}
