% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp2}
\alias{dgp2}
\title{Data generating process (summary statistics) with high dimensional exposure.}
\usage{
dgp2(ny, nx, beta_star, type = c("indep_mv", "depen_mv"), seed_num = NULL)
}
\description{
Data generating process (summary statistics) with high dimensional exposure.
}
