% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp3}
\alias{dgp3}
\title{DGP with 5 exposure and 5 instruments but 2 instruments}
\usage{
dgp3(n, beta1, beta2, d1 = 1, d2 = 2, seed_num = NULL)
}
\description{
DGP with 5 exposure and 5 instruments but 2 instruments
}
\details{
2 of the 5 exposure have a causal effect on the outcome.
}
