% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_all_valid}
\alias{dgp_all_valid}
\title{Data generating process 3}
\usage{
dgp_all_valid(n, beta1, beta2, seed_num = NULL)
}
\description{
Data generating process 3
}
\details{
DGP with 5 exposure and 5 instruments which are all
invalid, and 2 of the 5 exposure have a causal effect on the outcome.
}
