% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_ex0}
\alias{dgp_ex0}
\title{Data generating process 0 (valid & over-identified & correlated IVs)}
\usage{
dgp_ex0(n, beta_star, seed_num = NULL)
}
\arguments{
\item{beta_star}{A numeric vector of length 2.}
}
\description{
Data generating process 0 (valid & over-identified & correlated IVs)
}
\details{
An over-identified valid IV example with 4 correlated instruments
and 2 exposures.
}
