% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_ex00}
\alias{dgp_ex00}
\title{Data generating process 00 (2 valid & indep. IVs + 2 cond. indep. exposures)}
\usage{
dgp_ex00(n, beta_star = c(1, 2), seed_num = NULL)
}
\arguments{
\item{beta_star}{A numeric vector of length 2.}
}
\description{
Data generating process 00 (2 valid & indep. IVs + 2 cond. indep. exposures)
}
\details{
If beta_star does not contain 0, this is a dgp with 2 valid and
independent instruments, 2 exposures, and 2 independent hidden confounders
between exposures and outcome.
}
