% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_ex01}
\alias{dgp_ex01}
\title{Data generating process with 20 correlated instruments and 5 exposures.}
\usage{
dgp_ex01(n, beta1, beta2, d1 = 1, d2 = 1, seed_num = NULL)
}
\description{
Data generating process with 20 correlated instruments and 5 exposures.
}
\details{
If d1 and/or d2 are non-zero, the corresponding instruments are
invalid.
}
