% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_ex03_sumstat}
\alias{dgp_ex03_sumstat}
\title{Data generating process (summary statistics) with high dimensional exposure.}
\usage{
dgp_ex03_sumstat(
  ny,
  nx,
  beta_star,
  type = c("indep_mv", "depen_mv"),
  seed_num = NULL
)
}
\description{
Data generating process (summary statistics) with high dimensional exposure.
}
