% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_ex1}
\alias{dgp_ex1}
\title{Data generating process in Example 1 of spaceIV}
\usage{
dgp_ex1(n, beta2, seed_num = NULL)
}
\description{
Data generating process in Example 1 of spaceIV
}
\details{
Example 1 in spaceIV with three hidden variables each affecting one
exposure as well as the outcome. A DGP with 2 instruments and 3 exposures.
}
