% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{dgp_with_invalid}
\alias{dgp_with_invalid}
\title{Data generating process 2}
\usage{
dgp_with_invalid(n, beta1, beta2, d1 = 1, d2 = 2, seed_num = NULL)
}
\description{
Data generating process 2
}
\details{
DGP with 5 exposure and 5 instruments but 2 instruments
are invalid; 2 of the 5 exposure have a causal effect on the outcome.
}
