% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{get_block}
\alias{get_block}
\title{Extract the ij-th block of a squared block structured matrix}
\usage{
get_block(ii, jj, block_mat, dim_block)
}
\arguments{
\item{dim_block}{A scalar for the dimension of each block.}
}
\description{
Extract the ij-th block of a squared block structured matrix
}
