% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{get_submat}
\alias{get_submat}
\title{Extract a submatrix of a squared block structured matrix}
\usage{
get_submat(S, block_mat, dim_block)
}
\arguments{
\item{dim_block}{A scalar for the dimension of each block.}
}
\description{
Extract a submatrix of a squared block structured matrix
}
\details{
Extract the submatrix containing row and column blocks indexed
by elements in S.
}
