% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{grid_search_iv_lasso}
\alias{grid_search_iv_lasso}
\title{Grid search for IV-Lasso}
\usage{
grid_search_iv_lasso(
  beta_init,
  X,
  Y,
  Z,
  pen_grid = pracma::logspace(-3, 1, 30),
  alpha = 0.05,
  chisq_df = NULL,
  max_iter = 500,
  refit_set = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{beta_init}{A vector of initial value of beta.}

\item{pen_grid}{A vector of penalty parameter (lambda) values to go through.}

\item{alpha}{A scalar in (0, 1) for the significance level for the AR test.}

\item{max_iter}{An integer indicating the maximum number of iterations in
coordinate descent.}

\item{refit_set}{A boolean indicating whether to refit the model
after selecting the variables with Lasso.}

\item{verbose}{A boolean passed to \code{coordinate_descent_iv_lasso}
for whether to print messages.}
}
\description{
Grid search for IV-Lasso
}
