% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{grid_search_iv_lasso_suffstat}
\alias{grid_search_iv_lasso_suffstat}
\title{Grid search for IV-Lasso using sufficient statistics}
\usage{
grid_search_iv_lasso_suffstat(
  beta_init,
  XX,
  YY,
  ZZ,
  XY,
  ZX,
  ZY,
  nobs,
  pen_grid = pracma::logspace(-3, 1, 30),
  alpha = 0.05,
  chisq_df = NULL,
  max_iter = 500,
  abs_tol = 1e-05,
  refit_set = FALSE,
  verbose = FALSE
)
}
\description{
Grid search for IV-Lasso using sufficient statistics
}
