% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{grid_search_q_stat_gen_l1}
\alias{grid_search_q_stat_gen_l1}
\title{Grid search for Q-Lasso.}
\usage{
grid_search_q_stat_gen_l1(
  beta_init,
  gamma_res,
  sig_res,
  gamma_exp,
  sig_exp,
  pen_grid = pracma::logspace(-3, 1, 30),
  stop_if_accept = TRUE,
  stop_rule = c("max_pen", "min_pen", "max_pval"),
  alpha = 0.05,
  chisq_df = NULL,
  abs_tol = 1e-05,
  optim_method = "L-BFGS-B",
  control = list(maxit = 500, factr = 1e+06),
  refit_set = FALSE,
  verbose = FALSE,
  ...
)
}
\description{
Grid search for Q-Lasso.
}
