% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{iv2_estimator}
\alias{iv2_estimator}
\title{Minimizing ||Cov(Z, Y-Xb)||^2}
\usage{
iv2_estimator(X, Y, Z)
}
\description{
Minimizing ||Cov(Z, Y-Xb)||^2
}
\details{
IV solutions using valid instruments satisfy Cov(Z, Y-Xb) = 0.
The objective here is to minimize the squared LHS.
Cannot be used in under-identified case.
In just-identified case, this gives the same result as
\code{iv_estimator}.
}
\examples{
dat <- dgp_ex00(1e4, c(1, 2))
iv2_estimator(dat$X, dat$Y, dat$Z)
dat <- dgp_ex0(1e3, c(1, 2))
iv2_estimator(dat$X, dat$Y, dat$Z)

}
