% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{iv2_estimator_sumstat}
\alias{iv2_estimator_sumstat}
\title{Minimizing ||Cov(Z, Y-Xb)||^2 using summary statistics and variance of Z}
\usage{
iv2_estimator_sumstat(
  gamma_res,
  sig_res,
  gamma_exp,
  sig_exp,
  cov_zinx = NULL,
  cov_ziny = NULL,
  use_ginv = FALSE
)
}
\description{
Minimizing ||Cov(Z, Y-Xb)||^2 using summary statistics and variance of Z
}
\details{
IV solutions using valid instruments satisfy Cov(Z, Y-Xb) = 0.
The objective here is to minimize the squared LHS.
Cannot be used in under-identified case.
In just-identified case, this gives the same result as
\code{just_identified_iv_estimator}.
TODO: currently not using the sigs...
}
\examples{
dat <- dgp_ex00(1e4, c(1, 2))
dat_sumstat <- comp_sumstat(dat$Y, dat$Z, dat$X, dat$Z, type = "depen_mv")
iv2_estimator_sumstat(dat_sumstat$coef_y, dat_sumstat$cov_coef_y, 
                      dat_sumstat$coef_x, dat_sumstat$cov_coef_x, 
                      cov(dat$Z), cov(dat$Z))
iv2_estimator(dat$X, dat$Y, dat$Z)
dat <- dgp_ex0(1e3, c(1, 2))
dat_sumstat <- comp_sumstat(dat$Y, dat$Z, dat$X, dat$Z, type = "depen_mv")
iv2_estimator_sumstat(dat_sumstat$coef_y, dat_sumstat$cov_coef_y, 
                      dat_sumstat$coef_x, dat_sumstat$cov_coef_x, cov(dat$Z))
iv2_estimator(dat$X, dat$Y, dat$Z)

}
