% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{iv_estimator}
\alias{iv_estimator}
\title{IV estimator under just-identified scenario}
\usage{
iv_estimator(X, Y, Z)
}
\description{
IV estimator under just-identified scenario
}
\details{
The simplest just-identified IV estimator.
}
\examples{
dat <- dgp_ex00(1e3, c(1, 2))
iv_estimator(dat$X, dat$Y, dat$Z)

}
