% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{ivw_estimator}
\alias{ivw_estimator}
\title{Inverse-variance weighted estimator}
\usage{
ivw_estimator(
  gamma_res,
  sig_res,
  gamma_exp,
  sig_exp,
  weight_order = c("first", "second")
)
}
\arguments{
\item{weight_order}{One of "first" or "second" indicating whether to use
first or second order weights (inverse variance, which is approximated
using delta method).}
}
\description{
Inverse-variance weighted estimator
}
\details{
IVW assumes independent instruments.
}
\references{
Bowden, J., Del Greco M, F., Minelli, C., Zhao, Q.,
Lawlor, D.A., Sheehan, N.A., Thompson, J. and Davey Smith, G., 2019.
Improving the accuracy of two-sample summary-data Mendelian randomization:
moving beyond the NOME assumption. International journal of epidemiology,
48(3), pp.728-742.
}
