% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{ivw_weights}
\alias{ivw_weights}
\title{Calculate weights for the IVW estimator.}
\usage{
ivw_weights(gamma_res, sig_res)
}
\arguments{
\item{gamma_res}{A vector of length J, the estimated instrument-outcome effects.}

\item{sig_res}{A vector of length J, the variance of \code{gamma_res}.}
}
\description{
Calculate weights for the IVW estimator.
}
\details{
TODO: does this apply in the case of multiple exposures?
}
