% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-class_estimator.R
\name{k_class_suffstat}
\alias{k_class_suffstat}
\title{K-class estimator using sufficient statistics}
\usage{
k_class_suffstat(k, XX, ZZ, XY, ZX, ZY)
}
\arguments{
\item{k}{A scalar for the k-value.}

\item{XX}{A matrix of dimension R x R, the covariance matrix of the exposures.}

\item{ZZ}{A matrix of dimension J x J, the covariance matrix of the instruments.}

\item{XY}{A vector of length R, the covariances between the exposures and the response.}

\item{ZX}{A matrix of dimension J x R, the covariances between the instruments and the exposures.}

\item{ZY}{A vector of length J, the covariances between the instruments and the response.}
}
\description{
K-class estimator using sufficient statistics
}
\details{
Adapted from the code of spaceIV.
}
