% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-class_estimator.R
\name{liml_k}
\alias{liml_k}
\title{Calculate the k-value for LIML}
\usage{
liml_k(X, Y, P_Z, nobs)
}
\description{
Calculate the k-value for LIML
}
\details{
See e.g., https://en.wikipedia.org/wiki/Simultaneous_equations_model#Limited_information_maximum_likelihood_(LIML)
}
