% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{normal_approx_ci}
\alias{normal_approx_ci}
\title{Confidence intervals by normal approximation}
\usage{
normal_approx_ci(beta_hat, second_deriv, alpha = 0.05)
}
\description{
Confidence intervals by normal approximation
}
