% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{q_stat_gen_grad_k}
\alias{q_stat_gen_grad_k}
\title{Gradient of the general Q-statistic w.r.t beta_k}
\usage{
q_stat_gen_grad_k(kk, coef_y, cov_coef_y, coef_x, cov_coef_x, beta)
}
\description{
Gradient of the general Q-statistic w.r.t beta_k
}
