% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{q_stat_gen_l1}
\alias{q_stat_gen_l1}
\title{Subgradient of general Q-statistic with L1-penalty}
\usage{
q_stat_gen_l1(Y, sig_Y, X, sig_X, beta, lambda, tol = 1e-05)
}
\description{
Subgradient of general Q-statistic with L1-penalty
}
