% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{q_test}
\alias{q_test}
\title{Single-variable Q-test given independent instruments.}
\usage{
q_test(coef_y, se_coef_y, coef_x, se_coef_x, beta)
}
\arguments{
\item{coef_y}{A vector of length J, the estimated instrument-outcome coefficients.}

\item{se_coef_y}{A vector of length J, the standard error of the estimated instrument-outcome coefficients.}

\item{coef_x}{A vector of length J, the estimated instrument-exposure coefficients.}

\item{se_coef_x}{A vector of length J, the standard error of the estimated instrument-exposure coefficients.}

\item{beta}{A scalar, the exposure-outcome coefficient to be tested.}
}
\value{
A list containing the test statistic \code{stat} and the corresponding
p-value \code{pval}.
}
\description{
Single-variable Q-test given independent instruments.
}
