% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{q_test_gen}
\alias{q_test_gen}
\title{General Q-test that allows correlated instruments.}
\usage{
q_test_gen(
  coef_y,
  cov_coef_y,
  coef_x,
  cov_coef_x,
  beta,
  chisq_df = NULL,
  tol = 1e-05
)
}
\arguments{
\item{coef_y}{A vector of length J, the estimated instrument-outcome coefficients.}

\item{cov_coef_y}{A matrix of dimension J x J, the variance-covariance matrix of the estimated instrument-outcome coefficients.}

\item{coef_x}{A matrix of dimension J x K, the estimated instrument-exposure coefficients.}

\item{cov_coef_x}{A matrix of dimension (K x J) x (K x J) in a block form containing the variances and covariances of the estimated instrument-exposure coefficients. See patel-et-al2023 supplementary materials.}

\item{beta}{A vector of length K, the exposure-outcome coefficients.}

\item{tol}{A scalar, tolerance for considering an entry of beta to be zero.}
}
\value{
A list containing the test statistic \code{stat} and the corresponding
p-value \code{pval}.
}
\description{
General Q-test that allows correlated instruments.
}
