% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{q_test_gen_subvec_ci}
\alias{q_test_gen_subvec_ci}
\title{Subvector confidence interval with general Q test based on projection}
\usage{
q_test_gen_subvec_ci(
  coef_y,
  cov_coef_y,
  coef_x,
  cov_coef_x,
  beta_liml,
  subvec_idx = NULL,
  search_lwr = 1,
  search_upr = 1,
  extend = TRUE,
  alpha = 0.05,
  chisq_df = NULL,
  tol = 1e-05,
  uniroot_maxiter = 500,
  optim_method = "L-BFGS-B",
  control = list(maxit = 500, factr = 1e+06),
  ...
)
}
\description{
Subvector confidence interval with general Q test based on projection
}
