% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_tests.R
\name{q_test_mv}
\alias{q_test_mv}
\title{Multivariable Q-test given independent instruments.}
\usage{
q_test_mv(
  coef_y,
  se_coef_y,
  coef_x,
  cov_coef_x,
  beta,
  chisq_df = NULL,
  tol = 0.001
)
}
\arguments{
\item{coef_y}{A vector of length J, the estimated instrument-outcome coefficients.}

\item{se_coef_y}{A vector of length J, the standard error of the estimated instrument-outcome coefficients.}

\item{coef_x}{A matrix of dimension J x K, the estimated instrument-exposure coefficients.}

\item{cov_coef_x}{A list of J matrices each of dimension K x K, the covariance matrix of the estimated instrument-exposure coefficients.}

\item{beta}{A vector of length K, the exposure-outcome coefficient to be tested.}
}
\value{
p-value of the test.
}
\description{
Multivariable Q-test given independent instruments.
}
