% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{sandwich_ci}
\alias{sandwich_ci}
\title{Note: this implement w.r.t. ||Gamma - gamma \%*\% beta||^2 not general Q stat
since the first derivative of the general Q stat cannot be written
as a single sum of J terms. It seems that the sandwich estimator is
only studied under (quasi-)likelihood and generalized estimating
equations (GEEs) which both can be written as a single sum of
individual terms without interactions. Also not sure if the
sandwich brings more efficiency than using the variance of the
GMM estimator already.}
\usage{
sandwich_ci(coef_y, cov_coef_y, coef_x, cov_coef_x, beta_hat, alpha = 0.05)
}
\description{
Note: this implement w.r.t. ||Gamma - gamma \%*\% beta||^2 not general Q stat
since the first derivative of the general Q stat cannot be written
as a single sum of J terms. It seems that the sandwich estimator is
only studied under (quasi-)likelihood and generalized estimating
equations (GEEs) which both can be written as a single sum of
individual terms without interactions. Also not sure if the
sandwich brings more efficiency than using the variance of the
GMM estimator already.
}
