% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{spaceIV_estimator}
\alias{spaceIV_estimator}
\title{SpaceIV estimator (subset selection with Anderson-Rubin test)}
\usage{
spaceIV_estimator(
  X,
  Y,
  Z,
  max_size = NULL,
  alpha = 0.05,
  chisq_df = NULL,
  use_liml = TRUE,
  stop_if_accept = TRUE,
  stop_rule = c("min_set", "max_set", "max_pval")
)
}
\arguments{
\item{max_size}{The maximum size of subset to go through.}

\item{alpha}{A numeric scalar in (0,1), the significance level.}

\item{use_liml}{A boolean indicating whether to use LIML estimator to
estimate the coefficients for each subset.}

\item{stop_if_accept}{A boolean indicating whether to stop going through
larger subsets once AR test is not rejected.}
}
\description{
SpaceIV estimator (subset selection with Anderson-Rubin test)
}
