% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{spaceIV_estimator_inner}
\alias{spaceIV_estimator_inner}
\title{Inner SpaceIV estimator for a specific size of subsets}
\usage{
spaceIV_estimator_inner(X, Y, Z, size, use_liml = TRUE)
}
\arguments{
\item{use_liml}{A boolean indicating whether to use LIML estimator.}
}
\description{
Inner SpaceIV estimator for a specific size of subsets
}
\references{
Pfister and Peters 2022
}
