% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{spaceIV_estimator_suffstat}
\alias{spaceIV_estimator_suffstat}
\title{SpaceIV estimator using sufficient statistics}
\usage{
spaceIV_estimator_suffstat(
  XX,
  YY,
  ZZ,
  XY,
  ZX,
  ZY,
  nobs,
  max_size = NULL,
  alpha = 0.05,
  chisq_df = NULL,
  use_liml = FALSE
)
}
\description{
SpaceIV estimator using sufficient statistics
}
