% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{sparseMVMR_estimator}
\alias{sparseMVMR_estimator}
\title{Subset selection with Q-test}
\usage{
sparseMVMR_estimator(
  gamma_res,
  sig_res,
  gamma_exp,
  sig_exp,
  beta_init = NULL,
  indep_inst = TRUE,
  alpha = 0.05,
  chisq_df = NULL,
  max_size = NULL,
  stop_if_accept = TRUE,
  stop_rule = c("min_set", "max_set", "max_pval"),
  digits = 3,
  optim_method = "L-BFGS-B",
  control = list(factr = 1e+06),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{gamma_res}{A vector of length J, where J is the number of instruments.}

\item{sig_res}{If \code{indep_inst = TRUE}, a vector of length J, the standard
error of the estimated instrument-outcome coefficients; if \code{indep_inst = FALSE},
a matrix of dimension J x J, the variance-covariance matrix of the
estimated instrument-outcome coefficients.}

\item{gamma_exp}{A matrix of dimension J x K, where K is the number of exposures.}

\item{sig_exp}{A list of length J of matrices each of dimension K x K.}

\item{indep_inst}{A boolean indicating whether to assume the instruments are
independent. This affects the dimension of \code{sig_res}.}

\item{alpha}{A scalar indicating the confidence level.}

\item{chisq_df}{An integer to overwrite the degrees of freedom in the Q test.
if NULL, then J - number of non-zero entry of beta_hat will be used.}

\item{optim_method}{A string indicating the method used in \code{optim}.}

\item{control}{A list of contrl variables passing to \code{optim}. For example,
\code{list(reltol = sqrt(.Machine$double.eps))} if using "BFGS" method.}

\item{...}{Other parameters to pass on to \code{optim}. For example \code{lower} and
\code{upper} when using "L-BFGS-B" method.}
}
\description{
Subset selection with Q-test
}
