% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{sparseMVMR_estimator_inner}
\alias{sparseMVMR_estimator_inner}
\title{Fixed-size subsets optimization for the sparseMVMR estimator.}
\usage{
sparseMVMR_estimator_inner(
  gamma_res,
  sig_res,
  gamma_exp,
  sig_exp,
  size,
  beta_init = NULL,
  indep_inst = TRUE,
  chisq_df = NULL,
  abs_tol = 1e-05,
  optim_method = "L-BFGS-B",
  control = list(factr = 1e+06),
  ...
)
}
\arguments{
\item{gamma_res}{A vector of length J.}

\item{sig_res}{If \code{indep_inst = TRUE}, a vector of length J, the standard
error of the estimated instrument-outcome coefficients; if \code{indep_inst = FALSE},
a matrix of dimension J x J, the variance-covariance matrix of the
estimated instrument-outcome coefficients.}

\item{gamma_exp}{A matrix of dimension J x K.}

\item{sig_exp}{A list of variance-covariance matrices each of dimension J x K.}

\item{size}{A scalar indicating the size of the subsets of the exposures.}

\item{optim_method}{A string indicating the method used in \code{optim}.}

\item{control}{A list of contrl variables passing to \code{optim}. For example,
\code{list(reltol = sqrt(.Machine$double.eps))} if using "BFGS" method.}

\item{...}{Other parameters to pass on to \code{optim}. For example \code{lower} and
\code{upper} when using "L-BFGS-B" method.}
}
\description{
Fixed-size subsets optimization for the sparseMVMR estimator.
}
