% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sqrt_mat}
\alias{sqrt_mat}
\title{Take square-root of a matrix}
\usage{
sqrt_mat(X, tol = 1e-05)
}
\arguments{
\item{X}{A positive semi-definite matrix.}
}
\value{
"Square-root" of the input matrix.
}
\description{
Take square-root of a matrix
}
