% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{summarize_pvals}
\alias{summarize_pvals}
\title{Summarize p-values for subset selection}
\usage{
summarize_pvals(pval_all, var_names, digits = NULL)
}
\arguments{
\item{pval_all}{A list of lists each containing the p-values of sets
of one size.}
}
\description{
Summarize p-values for subset selection
}
