% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{tsiv_estimator}
\alias{tsiv_estimator}
\title{Two-sample IV estimator under just-identified scenario}
\usage{
tsiv_estimator(X, Z_X, Y, Z_Y)
}
\description{
Two-sample IV estimator under just-identified scenario
}
\details{
The simplest just-identified IV estimator.
}
\examples{
dat1 <- dgp_ex00(1e3, c(1, 2))
dat2 <- dgp_ex00(1e3, c(1, 2))
tsiv_estimator(dat1$X, dat1$Z, dat2$Y, dat2$Z)

}
