% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{tsls_estimator}
\alias{tsls_estimator}
\title{Two-stage least squares (TSLS) estimator}
\usage{
tsls_estimator(X, Y, Z)
}
\description{
Two-stage least squares (TSLS) estimator
}
\details{
Cannot be used in under-identified case.
In just-identified case, this gives the same result as
\code{iv_estimator}.
}
\examples{
dat <- dgp_ex00(1e3, c(1, 2))
tsls_estimator(dat$X, dat$Y, dat$Z)
dat <- dgp_ex0(1e3, c(1, 2))
tsls_estimator(dat$X, dat$Y, dat$Z)

}
