% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iv_estimators.R
\name{tstsls_estimator}
\alias{tstsls_estimator}
\title{Two-sample two-stage least-squares estimator}
\usage{
tstsls_estimator(X, Z_X, Y, Z_Y)
}
\description{
Two-sample two-stage least-squares estimator
}
\examples{
dat1 <- dgp_ex00(1e3, c(1, 2))
dat2 <- dgp_ex00(1e3, c(1, 2))
tstsls_estimator(dat1$X, dat1$Z, dat2$Y, dat2$Z)

}
